+++ bench/Benchmarks.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 
 using BenchmarkDotNet.Attributes;
 
+using OpaDotNet.Compilation.Abstractions;
 using OpaDotNet.Compilation.Cli;
 using OpaDotNet.Compilation.Interop;